cleanup: move wacz cert verification to certutils
